refactor imports and typings to satisfy mypy and sphinx
